Adds logging for debugging H264 stream reading
Adds logging statements to the H264 stream reader to aid in debugging issues related to finding start codes and reading data. This helps to understand when the reader is unable to find the next start code, how much data is being read, and when the end of the input stream is reached.